+++ src/main/java/org/jsoup/select/StructuralEvaluator.java
@@ -58,9 +58,11 @@ abstract class StructuralEvaluator extends Evaluator {
         }
     }
 
-    static class Has extends StructuralEvaluator {
-        final NodeIterator<Element> it = new NodeIterator<>(new Element("html"), Element.class);
+    final ThreadLocal<NodeIterator<Element>> threadHasIter =
+        ThreadLocal.withInitial(() -> new NodeIterator<>(new Element("html"), Element.class));
         // the element here is just a placeholder so this can be final - gets set in restart()
+
+    static class Has extends StructuralEvaluator {
 
         public Has(Evaluator evaluator) {
             super(evaluator);
@@ -68,6 +70,8 @@ abstract class StructuralEvaluator extends Evaluator {
 
         @Override public boolean matches(Element root, Element element) {
             // for :has, we only want to match children (or below), not the input element. And we want to minimize GCs
+            NodeIterator<Element> it = threadHasIter.get();
+
             it.restart(element);
             while (it.hasNext()) {
                 Element el = it.next();